requirements.txt
langchain
openai
streamlit
faiss-cpu
pypdf
python-dotenv
langchain-community
langchain-openai
